Replaced all instances of the vagrant user/passwords with alternatives.
In order to de-vagrantify the templates in this repository, we remove
all references to the vagrant user for authentication in favor of
more general usernames and passwords.

Essentially, the username and password pair, "vagrant", has been changed
to "user" for both usernames and passwords. The password for the
administrative user, "root", also had its password changed from
"vagrant" to "password". In response to these changes, any descriptions
that referenced the "vagrant" user, were reworded to "general user".

The documentation was updated to document these changes.