Added the --yes parameter to apt-get when installing libpam-systemd on debian/ubuntu.
The {debian,ubuntu}/systemd_{debian,ubuntu}.sh scripts under the
"packer_templates" directory use apt-get to install the "libpam-systemd"
package for the purpose of fixing an issue with ssh not being cleanly
terminated at https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=751636.
The intention of these systemd scripts is to do an unattended install of
the required packages, but due to a missing parameter it will block during
the build due to requiring a confirmation from the user to proceed.

To prevent apt-get from asking the user to confirm the installation of
libpam-systemd, the "--yes" parameter was added to both responsible
scripts. The "--yes" parameter is used by every other instance of
apt-get in this repository, so I believe that this is actually what the
author had originally intended.